requirements.txt
asgiref==3.9.1
azure-core==1.35.0
azure-storage-blob==12.25.1
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
coreapi==2.3.3
coreschema==0.0.4
cryptography==45.0.5
Django==5.2.4
django-cors-headers==4.7.0
django-storages==1.14.6
djangorestframework==3.16.0
idna==3.10
isodate==0.7.2
itypes==1.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
pillow==11.3.0
psycopg2-binary==2.9.10
pycparser==2.22
python-decouple==3.8
PyYAML==6.0.2
requests==2.32.4
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.14.1
uritemplate==4.2.0
urllib3==2.5.0
